50 % del proyecto

+++ app/Http/Controllers/ImagePostController.php
@@ -9,6 +9,8 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 use Mockery\Undefined;
+use FFMpeg\FFMpeg;
+use FFMpeg\Coordinate\TimeCode;
 
 class ImagePostController extends Controller
 {
@@ -16,21 +18,42 @@ class ImagePostController extends Controller
      * Display a listing of the resource.
      */
 
-     public function __construct()
-     {
-        // $this->middleware('auth')->except('index','show');
-     }
+    public function __construct()
+    {
+         $this->middleware('auth')->except('index','show');
+    }
 
 
     public function index()
     {
 
+        $user = Auth::user();
+
+        if ($user) {
+            $images = ImagePost::where(function ($query) use ($user) {
+                $query->where(function ($subQuery) use ($user) {
+                    $subQuery->where('private', 0)
+                        ->orWhere('user_post', $user->id);
+                });
+
+                if (!$user->pegi_18) {
+                    $query->where('pegi_18', false);
+                }
+            })->latest()->paginate(10);
+        } else {
+            $images = ImagePost::where('private', 0)
+                ->where('pegi_18', false)
+                ->latest()
+                ->paginate(10);
+        }
+
+        // $images = ImagePost::paginate(1);
 
 
-        $images = ImagePost::where('pegi_18' ,false)->orWhere('pegi_18',Auth::user()->pegi_18)->latest()->paginate(10)->onEachSide(1);
+        //  $images = ImagePost::where('pegi_18' ,false)->orWhere('pegi_18',Auth::user()->pegi_18)->latest()->paginate(10)->onEachSide(1);
 
 
-        return Inertia::render('images/ImagesList',compact('images',));
+        return Inertia::render('images/ImagesList', compact('images', ));
     }
 
     /**
@@ -47,38 +70,86 @@ class ImagePostController extends Controller
     public function store(Request $request)
     {
 
-        
-        $imagen = $request->image;
-        $imagenHash = hash_file('md5', $imagen->path());
-        $nombreImagen = uniqid().'.'.$imagen->getClientOriginalExtension();
-
         $image = new ImagePost();
+
+        $imagen = $request->image;
+        $extension = $imagen->getClientOriginalExtension();
+        $extensionesVideo = ['mp4', 'avi', 'mov', 'mkv', 'flv', 'wmv'];
+
+        if (in_array($extension, $extensionesVideo)) {
+            $videoPath = $imagen->path();
+            
+            $ffmpeg = FFMpeg::create([
+                'ffmpeg.binaries' => 'C:\Users\victo\OneDrive\Escritorio\dan\ffmpeg-6.0-essentials_build\ffmpeg-6.0-essentials_build\bin\ffmpeg.exe',
+                // Ruta a ffmpeg en tu sistema
+                'ffprobe.binaries' => 'C:\Users\victo\OneDrive\Escritorio\dan\ffmpeg-6.0-essentials_build\ffmpeg-6.0-essentials_build\bin\ffprobe.exe',
+                // Ruta a ffprobe en tu sistema
+                'timeout' => 3600,
+                'ffmpeg.threads' => 12,
+            ]);
+            $video = $ffmpeg->open($videoPath);
+
+
+            $duration = $video->getFormat()->get('duration');
+// Generar un número aleatorio dentro del rango de duración del video
+$randomTime = rand(1, $duration);
+
+
+            // Captura un fotograma del video en el segundo 5 (ajusta según tus necesidades)
+            $frame = $video->frame(TimeCode::fromSeconds($randomTime));
+
+            // Guarda el fotograma como imagen en storage
+            $lightVersionFilename = uniqid() . '.jpg';
+
+            // Guarda el fotograma como imagen en el almacenamiento (public disk)
+            $lightVersionPath =  $lightVersionFilename;
+            $frame->save(storage_path('app/public/light_versions/' . $lightVersionPath));
+
+            // Asigna la ubicación de la imagen generada al modelo ImagePost
+            $image->light_version_imagen = $lightVersionPath;
+
+
+
+
+
+        } else {
+            $imagenHash = hash_file('md5', $imagen->path());
+            $image->imagen_hash = $imagenHash;
+        }
+
+
+        $imagenHash = hash_file('md5', $imagen->path());
+        $nombreImagen = uniqid() . '.' . $extension;
+
+
         $image->name = $request->name;
         $image->original_url = $request->original_url;
         $image->danbooru_url = $request->danbooru_url;
         $image->user_post = Auth::user()->id;
         $image->pegi_18 = $request->pegi18;
+        $image->private = $request->private;
         $image->imagen = $nombreImagen;
-        $image->file_ext = $imagen->getClientOriginalExtension();
+        $image->file_ext = $extension;
         $image->file_size = $imagen->getSize();
-       
-       
-        $image->imagen_hash = $imagenHash; // aqui el hash
+
+
+        $image->imagen_hash = $imagenHash;
+        
         $image->save();
         Storage::disk('public')->putFileAs('imagesPost', $imagen, $nombreImagen);
 
-       
-
-        if($request->tags){
-            foreach($request->tags as $tag){
-                $image->tags()->attach($tag['value']);
-             }
-        } 
-        
 
 
+     if($request->tags){
+             foreach($request->tags as $tag){
+                 $image->tags()->attach($tag['value']);
+              }
+         } 
 
-        return to_route('images.show',$image);
+         dd('ok');
+
+
+        //return to_route('images.show', $image);
     }
 
     /**
@@ -87,8 +158,14 @@ class ImagePostController extends Controller
     public function show(ImagePost $image)
     {
 
-        if(($image->pegi_18 && !Auth::user()) || (!Auth::user()->pegi_18 && $image->pegi_18) ){
-            abort(403,'Acceso denegado');
+        if ($image->private) {
+            if (!auth() || Auth::user()->id != $image->user_post)
+                abort(403, 'Acceso denegado');
+
+        }
+
+        if (($image->pegi_18 && !Auth::user()) || (Auth::user() && !Auth::user()->pegi_18 && $image->pegi_18)) {
+            abort(403, 'Acceso denegado');
         }
 
         $tags = $image->tags;
@@ -101,20 +178,20 @@ class ImagePostController extends Controller
         $tag_artist = [];
         $tag_copyright = [];
         $tag_meta = [];
-        foreach($tags as $tag ){
-            if($tag->category  == 0){
-                $tag_general[] = $tag; 
-            } elseif($tag->category == 1 ){
+        foreach ($tags as $tag) {
+            if ($tag->category == 0) {
+                $tag_general[] = $tag;
+            } elseif ($tag->category == 1) {
                 $tag_copyright[] = $tag;
-            } elseif($tag->category == 2){
+            } elseif ($tag->category == 2) {
                 $tag_character[] = $tag;
-            } elseif($tag->category == 3){
+            } elseif ($tag->category == 3) {
                 $tag_artist[] = $tag;
-            }elseif($tag->category == 4) {
+            } elseif ($tag->category == 4) {
                 $tag_meta[] = $tag;
             }
         }
-        return Inertia::render('images/ImagesShow',compact('image','tag_general','tag_copyright','tag_character','tag_artist','tag_meta'));
+        return Inertia::render('images/ImagesShow', compact('image', 'tag_general', 'tag_copyright', 'tag_character', 'tag_artist', 'tag_meta'));
 
     }
 
@@ -127,17 +204,17 @@ class ImagePostController extends Controller
 
         $tag_list = [];
         foreach ($tags as $index => $tag) {
-            $tag_list [] = [
+            $tag_list[] = [
                 'value' => $tag->id,
                 'label' => $tag->name,
             ];
-            
-        }
-        
-       
-        return Inertia::render('images/FormImage',compact('image','tag_list'));
 
-        
+        }
+
+
+        return Inertia::render('images/FormImage', compact('image', 'tag_list'));
+
+
 
     }
 
@@ -146,7 +223,7 @@ class ImagePostController extends Controller
      */
     public function update(Request $request, ImagePost $image)
     {
-        
+
         //$image = ImagePost::findorFail($id);  
         $image->name = $request->name;
         $image->original_url = $request->original_url;
@@ -158,13 +235,13 @@ class ImagePostController extends Controller
 
         $image->tags()->detach();
 
-        if($request->tags){
-            foreach($request->tags as $tag){
+        if ($request->tags) {
+            foreach ($request->tags as $tag) {
                 $image->tags()->attach($tag['value']);
-             }
-        } 
+            }
+        }
 
-        return to_route('images.show',$image);
+        return to_route('images.show', $image);
 
     }
 
@@ -173,85 +250,93 @@ class ImagePostController extends Controller
      */
     public function destroy(string $id)
     {
-        $image = ImagePost::findorFail($id);   
-        
+        $image = ImagePost::findorFail($id);
+        if ($image->user_post != null && $image->user_post != Auth::user()->id) {
+            abort(403, 'solo el dueño puede borrarla');
+        }
+
+
         Storage::disk('public')->delete('imagesPost/' . $image->imagen);
+        Storage::disk('public')->delete('light_versions/' . $image->light_version_imagen);
+
+
         $image->delete();
-       
+
     }
 
- 
-
-    public function search(Request $request )
-{
-    $tag_id = $request->tags;
-    $tag_name = $request->tags_name;
 
 
-    $imagenesSearch = ImagePost::with('tags');
+    public function search(Request $request)
+    {
+        $tag_id = $request->tags;
+        $tag_name = $request->tags_name;
 
-    if($tag_id){
-        foreach ($tag_id as $tag) {        
-            $imagenesSearch->wherehas('tags', function ($q) use ($tag) {
-                $q->where('tag_id',$tag);
-            });
+
+        $imagenesSearch = ImagePost::with('tags');
+
+        if ($tag_id) {
+            foreach ($tag_id as $tag) {
+                $imagenesSearch->wherehas('tags', function ($q) use ($tag) {
+                    $q->where('tag_id', $tag);
+                });
+            }
+
         }
 
+        $tags = new Collection();
+        if ($tag_id && $tag_name) {
+            foreach ($tag_id as $index => $tag) {
+                $tags->push([
+                    'value' => $tag,
+                    'label' => $tag_name[$index],
+                ]);
+            }
+        }
+
+
+
+
+        if (!Auth::user() || (Auth::user() && !Auth::user()->pegi_18)) {
+            $imagenesSearch->where('pegi_18', '!=', true);
+        }
+
+        $images = $imagenesSearch->paginate(15);
+
+        $images->withQueryString();
+
+
+
+
+
+        return Inertia::render('images/ImagesList', compact('images', 'tags'));
     }
 
-    $tags = new Collection();
-    if ($tag_id && $tag_name) {
-        foreach ($tag_id as $index => $tag) {
-            $tags->push([
-                'value' => $tag,
-                'label' => $tag_name[$index],
-            ]);
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+    public function uploadByUrl(Request $request)
+    {
+
+
+        if ($request->url_search) {
+            //aqui el scraping a twitter
+        } else {
+
+            return Inertia::render('images/FormImageUrl');
+
         }
     }
-    
-
-
-
-    if(!Auth::user() || (Auth::user() && !Auth::user()->pegi_18) ){
-        $imagenesSearch->where('pegi_18','!=',true);
-    }
-
-    $images = $imagenesSearch->paginate(15);
-
-    $images->withQueryString();
-
-
-
-
-    
-    return Inertia::render('images/ImagesList',compact('images','tags'));
-}  
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-public function uploadByUrl(Request $request){
-
-
-    if($request->url_search){
-        //aqui el scraping a twitter
-    }else {
-
-    return Inertia::render('images/FormImageUrl');
-
-    }
-}
